Feat: rearranging sections so password field goes full width
On mobile, if graphic is on left, can't see full passwords on one line. Pulled password box out of main row and created a new row. Also added lots of comments identifying every open/close so I could tell what's what. Did a lot of Bootstrap col changes to make it look ok on mobile and on the big screen. Not perfect but pretty good.